// src/types/lunar-phase-card-config.ts
export type FontSizeOptions = 'auto' | 'small' | 'medium' | 'large' | 'x-large' | 'xx-large';
export type FontTextTransformOptions = 'none' | 'capitalize' | 'uppercase' | 'lowercase';
/**
 * @deprecated use 'font_config' instead
 */
export interface FontCustomStyles {
  header_font_size?: FontSizeOptions;
  header_font_style?: FontTextTransformOptions;
  header_font_color?: string;
  label_font_size?: FontSizeOptions;
  label_font_style?: FontTextTransformOptions;
  label_font_color?: string;
  hide_label?: boolean;
}